+++ src/pages/options/components/Footer/Footer.tsx
@@ -1,9 +1,19 @@
 import React from 'preact';
 import { preactTranslator } from '../../../../i18n';
 import { ResourceUrl } from '../../constants';
+import { Theme } from '../../../../theme';
+import { ThemeSwitch } from '../ThemeSwitch';
 import './footer.pcss';
 
-export const Footer: React.FunctionalComponent = () => (
+type FooterProps = {
+    theme: Theme,
+    toggleTheme: () => void,
+};
+
+export const Footer: React.FunctionalComponent<FooterProps> = ({
+    theme,
+    toggleTheme,
+}) => (
     <div class="footer">
         <div class="footer__in">
             <div class="footer__links-list">
@@ -14,6 +24,7 @@ export const Footer: React.FunctionalComponent = () => (
                     {preactTranslator.getMessage('homepage')}
                 </a>
             </div>
+            <ThemeSwitch theme={theme} toggleTheme={toggleTheme} />
         </div>
     </div>
 );